requirements.txt
Flask
flask-cors
opencv-python
numpy
imagehash
Pillow
matplotlib
scikit-image
torch
torchvision
web3
scikit-learn
